% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits_pe_rlc_data.R
\name{fits_pe_rlc_data}
\alias{fits_pe_rlc_data}
\title{Support tool for phytotools R package}
\usage{
fits_pe_rlc_data(
  data_id = NULL,
  data_par,
  data_fqfm,
  fit_methods = "Nelder-Mead",
  normalize = TRUE
)
}
\arguments{
\item{data_id}{Optional. By default NULL. Type "character" or "factor" expected. Vector of id(s) for a given RLC. If NULL, all data should be associated with a single RLC.}

\item{data_par}{Mandatory. Type "numeric" expected. Vector of PAR data. Units of umol m-2 s-1.}

\item{data_fqfm}{Mandatory. Type "numeric" expected. Vector of Photosynthetic rate or PSII quantum efficiency data.}

\item{fit_methods}{Mandatory. By default "Nelder-Mead". Type "character" or"factor" expected. You can use one of several methods among "Marq", "Port", "Newton", "Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN" or "Pseudo".}

\item{normalize}{Optional. By default TRUE. Type "boolean" expected. Set to TRUE if you want to normalize data.}
}
\value{
Return a list in the R environment.
}
\description{
This function provides an enhancement of the processes developed under the R package phytotools.
Initially, phytotools fits PE and RLC data to one of a four published PE models, Eilers and Peeters 1988, Jassby and Platt 1976, Platt, Gallegos and Harrison 1980 or Webb et al. 1974.
}
