% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_appender_tibble.r
\name{log_appender_tibble}
\alias{log_appender_tibble}
\title{Personalised log record appender function}
\usage{
log_appender_tibble(name_log_tibble = "log_storage")
}
\arguments{
\item{name_log_tibble}{Optional. Type "character" expected. Name of the tibble which will contain log information. Use the function log_storage to generate an empty tibble with a validated template.}
}
\value{
Return a tibble in the R environment.
}
\description{
This function is a personalised log record appender used according the function log_appender of the R package logger.
}
