% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_storage.r
\name{log_storage}
\alias{log_storage}
\title{Log tibble initialisation}
\usage{
log_storage()
}
\value{
Return a tibble in the R environment.
}
\description{
This function initialises a tibble with a specific format for log incrementation.
}
