% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark-package.R
\docType{package}
\name{spark-package}
\alias{spark}
\alias{spark-package}
\title{spark: Support Package for Analysis, Research and Knowledge}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package aims to provide support to our work through the availability of a several standardised functions for divers manipulations and analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/umr-marbec/spark}
  \item \url{https://umr-marbec.github.io/spark/}
  \item Report bugs at \url{https://github.com/umr-marbec/spark/issues}
}

}
\author{
\strong{Maintainer}: Mathieu Depetris \email{mathieu.depetris@cnrs.fr} (\href{https://orcid.org/0000-0001-8080-0531}{ORCID})

Other contributors:
\itemize{
  \item UMR MARBEC MARine Biodiversity Exploitation & Conservation [copyright holder]
}

}
\keyword{internal}
