% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparck-package.R
\docType{package}
\name{sparck-package}
\alias{sparck}
\alias{sparck-package}
\title{sparck: Support Package for Analysis, Research, Collaboration and Knowledge}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A system to provide support to our work through the availability of a several standardised functions for divers manipulations and analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/umr-marbec/sparck}
  \item \url{https://umr-marbec.github.io/sparck/}
  \item Report bugs at \url{https://github.com/umr-marbec/sparck/issues}
}

}
\author{
\strong{Maintainer}: Mathieu Depetris \email{mathieu.depetris@cnrs.fr} (\href{https://orcid.org/0000-0001-8080-0531}{ORCID})

Other contributors:
\itemize{
  \item UMR MARBEC MARine Biodiversity Exploitation & Conservation [copyright holder]
}

}
\keyword{internal}
